refactor : condition 패키지 제거
